+++ apps/openci_runner/lib/src/features/run.dart
@@ -1,4 +1,5 @@
 import 'package:dartssh2/dartssh2.dart';
+import 'package:openci_models/openci_models.dart';
 import 'package:openci_runner/src/commands/runner_command.dart';
 import 'package:openci_runner/src/service/dartssh2/dartssh2_service.dart';
 import 'package:openci_runner/src/service/logger_service.dart';
@@ -7,20 +8,28 @@ Future<void> runCommand({
   required SSHClient client,
   required String command,
   required String? currentWorkingDirectory,
+  required String jobId,
 }) async {
   final dartSSH2Service = dartSSH2ServiceSignal.value;
   final firestore = firestoreSignal.value!;
   final logger = loggerSignal.value;
+  final ref = firestore
+      .collection(buildJobsCollectionPath)
+      .doc(jobId)
+      .collection('logs')
+      .doc();
   await dartSSH2Service.executeCommand(
     client: client,
     command: command,
     currentWorkingDirectory: currentWorkingDirectory,
     onDoneError: (result) async {
-      final ref = firestore.collection('logsV2').doc();
-      await ref.set({
-        'log': result.stderr,
-        'createdAt': DateTime.now(),
-      });
+      final log = CommandLog(
+        command: command,
+        log: result.stderr,
+        createdAt: DateTime.now(),
+      );
+
+      await ref.set(log.toJson());
       logger.err(
         'Command failed: stdout: ${result.stdout}, stderr: ${result.stderr}',
       );
@@ -29,11 +38,13 @@ Future<void> runCommand({
       );
     },
     onDoneSuccess: (result) async {
-      final ref = firestore.collection('logsV2').doc();
-      await ref.set({
-        'log': result.stderr,
-        'createdAt': DateTime.now(),
-      });
+      final log = CommandLog(
+        command: command,
+        log: result.stderr,
+        createdAt: DateTime.now(),
+      );
+
+      await ref.set(log.toJson());
       logger.info(
         'Command executed successfully: stdout: ${result.stdout}, stderr: ${result.stderr}',
       );