+++ apps/openci_runner/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.8.7
+
+- fix: command log
+
+## 0.8.6
+
+- fix: command log
+
 ## 0.8.5
 
 - fix: workflow runner